Create a superclass GobiertoCalendars::EventForm to abstract two others
