+++ app/forms/gobierto_people/person_event_form.rb
@@ -1,67 +1,29 @@
 # frozen_string_literal: true
 
 module GobiertoPeople
-  class PersonEventForm < BaseForm
+  class PersonEventForm < ::GobiertoCalendars::EventForm
 
     prepend ::GobiertoCommon::Trackable
 
     attr_accessor(
-      :external_id,
-      :site_id,
-      :person_id,
       :title,
-      :description,
-      :starts_at,
-      :ends_at,
-      :state,
-      :notify,
-      :locations,
-      :attendees
+      :description
     )
 
-    delegate :persisted?, to: :person_event
-
-    validates :external_id, :title, :starts_at, :ends_at, :site, :collection, presence: true
+    validates :external_id, :title, presence: true
     validate :event_in_sync_range_window
 
-    trackable_on :person_event
-
+    trackable_on :event
     notify_changed :state
 
     def initialize(options = {})
-        ordered_options = HashWithIndifferentAccess.new(site_id: options[:site_id], person_id: options[:person_id])
-        ordered_options.merge!(options)
-        super(ordered_options)
-      end
-
-    def destroy
-      unless person_event.new_record?
-        person_event.destroy
-      end
+      ordered_options = HashWithIndifferentAccess.new(site_id: options[:site_id], person_id: options[:person_id])
+      ordered_options.merge!(options)
+      super(ordered_options)
     end
 
     def save
-      save_person_event if valid?
-    end
-
-    def site
-      @site ||= Site.find_by(id: site_id)
-    end
-
-    def person_event
-      @person_event ||= person.events.find_by(external_id: external_id).presence || build_person_event
-    end
-
-    def person
-      @person ||= site.people.find_by(id: person_id)
-    end
-
-    def collection
-      @collection ||= person.try(:events_collection)
-    end
-
-    def locations
-      @locations ||= []
+      save_event if valid?
     end
 
     def locations_attributes=(attributes)
@@ -70,28 +32,11 @@ module GobiertoPeople
       attributes.each do |_, location_attributes|
         next if location_attributes["_destroy"] == "1"
 
-        location = person_event.locations.detect{ |l| l.name == location_attributes[:name] } ||
-          person_event_location_class.new(name: location_attributes[:name])
+        location = event.locations.detect{ |l| l.name == location_attributes[:name] } ||
+          event_location_class.new(name: location_attributes[:name])
 
         @locations.push(location) if location.valid?
       end
-    end
-
-    def attendees
-      @attendees ||= person_event.attendees.presence || []
-
-      return @attendees if person.nil?
-
-      if person_event.attendees.any?
-        unless organizer = person_event.attendees.find_by(person_id: person_id)
-          organizer = person_event_attendee_class.new(person_id: person_id)
-        end
-      else
-        organizer = person_event_attendee_class.new(person_id: person_id)
-      end
-      @attendees.push(organizer) unless @attendees.any?{ |a| a.person_id == organizer.person_id }
-
-      @attendees
     end
 
     def attendees=(attendees_attributes)
@@ -103,16 +48,16 @@ module GobiertoPeople
 
         attendee_person = attendee_attributes[:person] || site.people.find_by(email: email)
 
-        existing_attendee = person_event.attendees.detect do |a|
+        existing_attendee = event.attendees.detect do |a|
           (attendee_person.present? && a.person == attendee_person) || a.name == name
         end
 
         attendee = if existing_attendee
                      existing_attendee
                    elsif attendee_person.present?
-                     person_event_attendee_class.new(person: attendee_person)
+                     event_attendee_class.new(person: attendee_person)
                    else
-                     person_event_attendee_class.new(name: name)
+                     event_attendee_class.new(name: name)
                    end
 
         @attendees.push(attendee) if attendee.valid?
@@ -120,30 +65,14 @@ module GobiertoPeople
     end
 
     def state
-      @state ||= person_event.state || "published"
+      @state ||= event.state || "published"
     end
 
     def notify?
-      notify && person_event.active? && person.present?
+      notify && event.active? && person.present?
     end
 
     private
-
-    def build_person_event
-      event_class.new site_id: site_id
-    end
-
-    def event_class
-      ::GobiertoCalendars::Event
-    end
-
-    def person_event_location_class
-      ::GobiertoCalendars::EventLocation
-    end
-
-    def person_event_attendee_class
-      ::GobiertoCalendars::EventAttendee
-    end
 
     def event_in_sync_range_window
       errors.add(:starts_at, 'is out of sync range') unless GobiertoCalendars.sync_range.cover?(starts_at)
@@ -163,38 +92,26 @@ module GobiertoPeople
       Hash[I18n.available_locales.map {|x| [x, nil]}]
     end
 
-    def save_person_event
-      @person_event = person_event.tap do |person_event_attributes|
-        person_event_attributes.site_id = site_id
-        person_event_attributes.external_id = external_id
-        person_event_attributes.collection = collection
-        person_event_attributes.state = state
-        person_event_attributes.title_translations = title_translations
-        person_event_attributes.description_translations = description_translations
-        person_event_attributes.starts_at = starts_at
-        person_event_attributes.ends_at = ends_at
-        person_event_attributes.locations = locations
-        person_event_attributes.attendees = attendees
-      end
+    def save_event
+      assign_event_attributes
 
-      if @person_event.valid?
-        if @person_event.changes.any?
+      if @event.valid?
+        if @event.changes.any?
           if notify?
             run_callbacks(:save) do
-              @person_event.save
+              @event.save
             end
           else
-            @person_event.save
+            @event.save
           end
         end
 
-        @person_event
+        @event
       else
-        promote_errors(@person_event.errors)
+        promote_errors(@event.errors)
 
         false
       end
     end
-
   end
 end
